Partly removed ArrayList. Writing now to wostream.

+++ source/applications/coco-r/include/CharClass.h
@@ -37,7 +37,7 @@ namespace Coco
 class CharClass
 {
 public:
-	int n;          // class number
+	size_t n;           // class number
 	std::wstring name;  // class name
 	CharSet *set;   // set representing the class
